Move api pagination tests
