+++ saleor/graphql/account/tests/test_pagination.py
@@ -3,8 +3,7 @@ from django.contrib.auth import models as auth_models
 
 from saleor.account.models import User
 from saleor.order.models import Order
-
-from ..utils import get_graphql_content
+from tests.api.utils import get_graphql_content
 
 
 @pytest.fixture()
