+++ saleor/graphql/menu/tests/test_pagination.py
@@ -1,8 +1,7 @@
 import pytest
 
 from saleor.menu.models import Menu, MenuItem
-
-from ..utils import get_graphql_content
+from tests.api.utils import get_graphql_content
 
 
 @pytest.fixture
