+++ saleor/graphql/order/tests/test_pagination.py
@@ -8,8 +8,7 @@ from prices import Money, TaxedMoney
 
 from saleor.order.models import Order, OrderStatus
 from saleor.payment import ChargeStatus
-
-from ..utils import get_graphql_content
+from tests.api.utils import get_graphql_content
 
 
 @pytest.fixture()
